requirements-test.txt
pip>=19.3.1
setuptools>=42
pytest==6.2.5
pytest-django==4.5.2
pretend
watchdog
libsass
coverage
pytest-cov==4.0.0
tox
mypy
django-stubs[compatible-mypy]
